Rename to .css since we're not using any SCSS
