+++ src/renderer/components/SearchInput/styles.module.css
